Move caddy base config test to the config tests
